+++ juniper/src/macros/tests/impl_object.rs
@@ -35,7 +35,7 @@ struct Query {
 
 #[crate::object_internal(
     scalar = crate::DefaultScalarValue,
-    name = "Query", 
+    name = "Query",
     context = Context,
 )]
 /// Query Description.
@@ -84,14 +84,16 @@ impl<'a> Query {
         arg1
     }
 
-    #[graphql(arguments(default_arg(default = true)))]
-    fn default_argument(default_arg: bool) -> bool {
+    fn default_argument(#[graphql(default = true)] default_arg: bool) -> bool {
         default_arg
     }
 
-    #[graphql(arguments(arg(description = "my argument description")))]
-    fn arg_with_description(arg: bool) -> bool {
+    fn arg_with_description(#[graphql(description = "my argument description")] arg: bool) -> bool {
         arg
+    }
+
+    fn renamed_argument(#[graphql(name = new_name)] old_name: bool) -> bool {
+        old_name
     }
 
     fn with_context_child(&self) -> WithContext {
@@ -120,12 +122,19 @@ impl Mutation {
     }
 }
 
+fn juniper_value_to_serde_json_value(
+    value: &crate::Value<crate::DefaultScalarValue>,
+) -> serde_json::Value {
+    serde_json::from_str(&serde_json::to_string(value).unwrap()).unwrap()
+}
+
 #[test]
 fn object_introspect() {
     let res = util::run_info_query::<Query, Mutation, Context>("Query");
-    assert_eq!(
-        res,
-        crate::graphql_value!({
+
+    assert_json_diff::assert_json_include!(
+        actual: juniper_value_to_serde_json_value(&res),
+        expected: serde_json::json!({
             "name": "Query",
             "description": "Query Description.",
             "fields": [
@@ -136,32 +145,32 @@ fn object_introspect() {
                 },
                 {
                     "name": "independent",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [],
                 },
                 {
                     "name": "withExecutor",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [],
                 },
                 {
                     "name": "withExecutorAndSelf",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [],
                 },
                 {
                     "name": "withContext",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [],
                 },
                 {
                     "name": "withContextAndSelf",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [],
                 },
                 {
                     "name": "renamed",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [],
                 },
                 {
@@ -176,24 +185,24 @@ fn object_introspect() {
                 },
                 {
                     "name": "hasArgument",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [
                         {
                             "name": "arg1",
-                            "description": None,
+                            "description": None::<String>,
                             "type": {
-                                "name": None,
+                                "name": None::<String>,
                             },
                         }
                     ],
                 },
                 {
                     "name": "defaultArgument",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [
                         {
                             "name": "defaultArg",
-                            "description": None,
+                            "description": None::<String>,
                             "type": {
                                 "name": "Boolean",
                             },
@@ -202,36 +211,49 @@ fn object_introspect() {
                 },
                 {
                     "name": "argWithDescription",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [
                         {
                             "name": "arg",
                             "description": "my argument description",
                             "type": {
-                                "name": None
+                                "name": None::<String>
+                            },
+                        }
+                    ],
+                },
+                {
+                    "name": "renamedArgument",
+                    "description": None::<String>,
+                    "args": [
+                        {
+                            "name": "newName",
+                            "description": None::<String>,
+                            "type": {
+                                "name": None::<String>
                             },
                         }
                     ],
                 },
                 {
                     "name": "withContextChild",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [],
                 },
                 {
                     "name": "withLifetimeChild",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [],
                 },
                 {
                     "name": "withMutArg",
-                    "description": None,
+                    "description": None::<String>,
                     "args": [
                         {
                             "name": "arg",
-                            "description": None,
+                            "description": None::<String>,
                             "type": {
-                                "name": None,
+                                "name": None::<String>,
                             },
                         }
                     ],