+++ juniper_codegen/src/impl_object.rs
@@ -1,5 +1,6 @@
 use crate::util;
 use proc_macro::TokenStream;
+use proc_macro_error::*;
 use quote::quote;
 
 /// Generate code for the juniper::object macro.
@@ -101,14 +102,16 @@ pub fn build_object(args: TokenStream, body: TokenStream, is_internal: bool) -> 
                     }
                 };
 
-                let attrs = match util::FieldAttributes::from_attrs(
+                let mut attrs = match util::FieldAttributes::from_attrs(
                     method.attrs,
                     util::FieldAttributeParseMode::Impl,
                 ) {
                     Ok(attrs) => attrs,
-                    Err(err) => panic!(
+                    Err(err) => abort!(
+                        err.span(),
                         "Invalid #[graphql(...)] attribute on field {}:\n{}",
-                        method.sig.ident, err
+                        method.sig.ident,
+                        err
                     ),
                 };
 
@@ -134,7 +137,14 @@ pub fn build_object(args: TokenStream, body: TokenStream, is_internal: bool) -> 
                                     panic!("Invalid token for function argument");
                                 }
                             };
-                            let arg_name = arg_ident.to_string();
+                            let arg_ident_name = arg_ident.to_string();
+
+                            if let Some(field_arg) = util::parse_argument_attrs(&captured) {
+                                // We insert with `arg_ident_name` as the key because the argument
+                                // might have been renamed in the param attribute and we need to
+                                // look it up for making `final_name` further down.
+                                attrs.arguments.insert(arg_ident_name.clone(), field_arg);
+                            }
 
                             let context_type = definition.context.as_ref();
 
@@ -169,8 +179,13 @@ pub fn build_object(args: TokenStream, body: TokenStream, is_internal: bool) -> 
                                 // Regular argument.
 
                                 let ty = &captured.ty;
-                                // TODO: respect graphql attribute overwrite.
-                                let final_name = util::to_camel_case(&arg_name);
+
+                                let final_name = attrs
+                                    .argument(&arg_ident_name)
+                                    .and_then(|arg| arg.name.as_ref())
+                                    .map(|name| util::to_camel_case(&name.to_string()))
+                                    .unwrap_or_else(|| util::to_camel_case(&arg_ident_name));
+
                                 let expect_text = format!("Internal error: missing argument {} - validation must have failed", &final_name);
                                 let mut_modifier = if is_mut { quote!(mut) } else { quote!() };
                                 resolve_parts.push(quote!(
@@ -179,11 +194,11 @@ pub fn build_object(args: TokenStream, body: TokenStream, is_internal: bool) -> 
                                         .expect(#expect_text);
                                 ));
                                 args.push(util::GraphQLTypeDefinitionFieldArg {
-                                    description: attrs.argument(&arg_name).and_then(|arg| {
+                                    description: attrs.argument(&arg_ident_name).and_then(|arg| {
                                         arg.description.as_ref().map(|d| d.value())
                                     }),
                                     default: attrs
-                                        .argument(&arg_name)
+                                        .argument(&arg_ident_name)
                                         .and_then(|arg| arg.default.clone()),
                                     _type: ty.clone(),
                                     name: final_name,